stupid solution

+++ ProblemParams.cs
@@ -2,11 +2,11 @@ namespace _2021
 {
     internal class ProblemParams
     {
-        public short AllTime { get; internal set; }
-        public short IntersecCount { get; internal set; }
-        public short StreetCount { get; internal set; }
-        public short CarCount { get; internal set; }
-        public short Bonus { get; internal set; }
+        public int AllTime { get; internal set; }
+        public int IntersecCount { get; internal set; }
+        public int StreetCount { get; internal set; }
+        public int CarCount { get; internal set; }
+        public int Bonus { get; internal set; }
         public int DefaultTimeForLight { get; internal set; }
     }
 }
